requirements.txt
annotated-types==0.7.0
anyio==4.4.0
assemblyai==0.33.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
elevenlabs==1.7.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.38
langchain-openai==0.1.23
langsmith==0.1.114
openai==1.43.0
orjson==3.10.7
packaging==24.1
PyAudio==0.2.14
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
websockets==13.0.1
